fix user filter operations
Two functions implemented, one re-written, and two deleted:

1. posix_evfilt_user_knote_create and delete functions are implemented.
2. posix_evfilt_user_copyout is re-written such that it only copies
knote->kev back to user. This posix_kevent_copyout will provide the
knote.
3. posix_evfilt_user_init and posix_evfilt_user_destroy are deleted because
they are not needed anymore.

A pipe is created when creating a knote. The write end of pipe is
responsible to trigger event, while the read end of pipe is add to
kqueue's file descriptor list to be monitored.