use hash map to find knote with a given file descriptor
This patch uses a hash table with a custom memory allocator to find a
knote with a given file descriptor when copyout a triggered kevent.
Previous implementation uses an array of length MAX_FILE_DESCRIPTORS but
on z/OS the MAX_FILE_DESCRIPTORS is such a large number that knote_map
wastes too much memory.